Allow MessageBox to work without window
